requirements.txt
asgiref==3.5.2
attrs==22.2.0
boto3==1.26.74
botocore==1.29.74
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==3.2.13
django-cors-headers==3.10.0
django-crontab==0.7.1
django-mysql==4.9.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.21.4
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
model-bakery==1.10.1
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.0
pandas==2.0.2
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
